@@ -6,10 +6,10 @@
     "react-scripts": "^1.1.4"
   },
   "dependencies": {
-    "react": "^16.4.1",
-    "react-dom": "^16.4.1",
-    "react-redux": "^5.0.7",
-    "redux": "^4.0.0"
+    "react": "^16.8.0",
+    "react-dom": "^16.12.0",
+    "react-redux": "^7.1",
+    "redux": "^4.0.4"
   },
   "scripts": {
     "start": "react-scripts start",
