@@ -1,15 +1,18 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './App';
-import manageTodo from './reducers/manageTodo';
-import { Provider } from 'react-redux';
-import { createStore } from 'redux';
+import React from "react";
+import ReactDOM from "react-dom";
+import App from "./App";
+import manageTodo from "./reducers/manageTodo";
+import { Provider } from "react-redux";
+import { createStore } from "redux";
 
-let store = createStore(manageTodo);
+let store = createStore(
+  manageTodo,
+  window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
+);
 
 ReactDOM.render(
   <Provider store={store}>
     <App />
   </Provider>,
-  document.getElementById('root')
+  document.getElementById("root")
 );
