@@ -1,8 +1,14 @@
-export default function manageTodo(state = {
-  todos: [],
-}, action) {
-
+export default function manageTodo(
+  state = {
+    todos: []
+  },
+  action
+) {
   console.log("reducer received this action:", action);
-
-  return state;
+  switch (action.type) {
+    case "ADD_TODO":
+      return { ...state, todos: [...state.todos, action.payload] };
+    default:
+      return state;
+  }
 }
